Post-merge changes as a result of project reconstruction
