+++ Sources/Hyperspace/Service/Network/NetworkActivityIndicatable.swift
