+++ Tests/Helper/Mocks/MockNetworkActivityIndicator.swift
